logicport: Submission of working application (v0.1)
This is the beginnings of what will be the driver for the Introix
Logicport LA1034, which is a 34 channel logic analyser capable of
a 500MHz sample speed. This driver will be submitted to the Open
Source sigrok signal analysis project (http://sigrok.org).

This submission contains the basic application capable of
bringing up the hardware and sending out the first device control
packet.

NB: You will notice the commented out function calls during
start-up. These were left in intentionally for reverting purposes
and will be removed in the next commit.

Signed-off-by: Lee Jones <lee.jones@linaro.org>